Edit and delete branches

+++ colmanInternetStav1.1/Controllers/BranchesController.cs
@@ -8,6 +8,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using colmanInternetStav1._1.Models;
+using System.Security.Claims;
 using System.Net.Http;
 using Newtonsoft.Json;
 
@@ -15,32 +16,30 @@ namespace colmanInternetStav1._1.Controllers
 {
     public class BranchesController : Controller
     {
-        public struct Coordinates
-        {
-            public int lat, lng;
-
-            public Coordinates(int Lat, int Lng)
-            {
-                lat = Lat;
-                lng = Lng;
-            }
-        }
-
-        private Coordinates getCoordinatesFromAddressAsync(string address)
-        {
-            Coordinates coords;
-
-            coords.lat = 0;
-            coords.lng = 0;
-
-            return coords;
-        }
-
         private readonly ColmanInternetiotContext _context;
+
+        public bool isStoreOwner(ClaimsPrincipal principal, int storeId)
+        {
+            if (Account.isLoggedIn(User))
+            {
+                ColmanInternetiotContext context = new ColmanInternetiotContext();
+
+                int userId = context.Users.FirstOrDefault(c => c.NameId == Account.getDetails(principal)["nameid"]).Id;
+
+                int? managerId = context.Branch.FirstOrDefault(c => c.Id == storeId).ManagerId;
+
+                if (managerId != null)
+                {
+                    return (userId == managerId);
+                }
+            }
+
+            return false;
+        }
 
         public BranchesController(ColmanInternetiotContext context)
         {
-            _context = context;    
+            _context = context;
         }
 
         // GET: Branches
@@ -102,8 +101,6 @@ namespace colmanInternetStav1._1.Controllers
 
                 ViewData["ManagerId"] = new SelectList(_context.Users, "Id", "Id", branch.ManagerId);
 
-                Coordinates coords = getCoordinatesFromAddressAsync(branch.Address);
-
                 return View(branch);
             }
             else
@@ -120,13 +117,23 @@ namespace colmanInternetStav1._1.Controllers
                 return NotFound();
             }
 
-            var branch = await _context.Branch.SingleOrDefaultAsync(m => m.Id == id);
-            if (branch == null)
+            if (Account.isAdmin(User) || isStoreOwner(User, int.Parse(id.ToString())))
             {
-                return NotFound();
+                var branch = await _context.Branch.SingleOrDefaultAsync(m => m.Id == id);
+                if (branch == null)
+                {
+                    return NotFound();
+                }
+
+                ViewData["ManagerName"] = _context.Users.FirstOrDefault(c => c.Id == _context.Branch.FirstOrDefault(x => x.Id == int.Parse(id.ToString())).ManagerId).Name;
+
+                ViewData["ManagerId"] = new SelectList(_context.Users, "Id", "Name", branch.ManagerId);
+                return View(branch);
             }
-            ViewData["ManagerId"] = new SelectList(_context.Users, "Id", "Id", branch.ManagerId);
-            return View(branch);
+            else
+            {
+                return (new RedirectToActionResult("NotAuthorized", "Home", null));
+            }
         }
 
         // POST: Branches/Edit/5
@@ -141,28 +148,36 @@ namespace colmanInternetStav1._1.Controllers
                 return NotFound();
             }
 
-            if (ModelState.IsValid)
+            if (Account.isAdmin(User) || isStoreOwner(User, id))
             {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(branch);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!BranchExists(branch.Id))
+                    try
                     {
-                        return NotFound();
+                        _context.Update(branch);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException)
                     {
-                        throw;
+                        if (!BranchExists(branch.Id))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+                    return RedirectToAction("Index");
                 }
-                return RedirectToAction("Index");
+
+                ViewData["ManagerId"] = new SelectList(_context.Users, "Id", "Id", branch.ManagerId);
+                return View(branch);
             }
-            ViewData["ManagerId"] = new SelectList(_context.Users, "Id", "Id", branch.ManagerId);
-            return View(branch);
+            else
+            {
+                return (new RedirectToActionResult("NotAuthorized", "Home", null));
+            }
         }
 
         // GET: Branches/Delete/5
@@ -173,15 +188,25 @@ namespace colmanInternetStav1._1.Controllers
                 return NotFound();
             }
 
-            var branch = await _context.Branch
-                .Include(b => b.Manager)
-                .SingleOrDefaultAsync(m => m.Id == id);
-            if (branch == null)
+            if (Account.isAdmin(User) || isStoreOwner(User, int.Parse(id.ToString())))
             {
-                return NotFound();
+                ViewData["StoreName"] = _context.Branch.FirstOrDefault(c => c.Id == int.Parse(id.ToString())).Name;
+
+                var branch = await _context.Branch
+            .Include(b => b.Manager)
+            .SingleOrDefaultAsync(m => m.Id == id);
+                if (branch == null)
+                {
+                    return NotFound();
+                }
+
+                return View(branch);
+            }
+            else
+            {
+                return (new RedirectToActionResult("NotAuthorized", "Home", null));
             }
 
-            return View(branch);
         }
 
         // POST: Branches/Delete/5
@@ -189,10 +214,17 @@ namespace colmanInternetStav1._1.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var branch = await _context.Branch.SingleOrDefaultAsync(m => m.Id == id);
-            _context.Branch.Remove(branch);
-            await _context.SaveChangesAsync();
-            return RedirectToAction("Index");
+            if (Account.isAdmin(User) || isStoreOwner(User, id))
+            {
+                var branch = await _context.Branch.SingleOrDefaultAsync(m => m.Id == id);
+                _context.Branch.Remove(branch);
+                await _context.SaveChangesAsync();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return (new RedirectToActionResult("NotAuthorized", "Home", null));
+            }
         }
 
         private bool BranchExists(int id)
